+++ authentication/app/services/authenticationService.js
@@ -1,11 +1,13 @@
 /**
- * Le service contenant les requêtes liées à l'authentification.
+ * Le fichier contenant les requêtes liées à l'authentification.
  * @author GAURE Warren
  * @version 1.0
 */
 
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcryptjs');
+const os = require('os');
+const osUtils = require('os-utils');
 const User = require('../models/userModel');
 
 /**
@@ -59,6 +61,7 @@ const createRestaurateur = async (email, password, userType, phoneNumber) => {
             password: password,
             phoneNumber: phoneNumber,
             userType: userType,
+            referralCode: (Math.random() + 1).toString(36).substring(2),
             isSuspended: false
         });
 
@@ -154,6 +157,48 @@ const generateJWT = (userID, userType) => {
     return token;
 };
 
+/**
+ * Fonction permettant de récupérer les métriques de performance de l'application, à savoir :
+ * - CPU usage (usage du CPU)
+ * - Total memory (mémoire totale)
+ * - Free memory (mémoire inutilisée)
+ * - Used memory (mémoire utilisée)
+ * - Elapsed time (temps de réponse)
+ * @returns {object} Un objet contenant les métriques de performance de l'application.
+*/
+const getPerformanceMetrics = async () => {
+    const startTime = Date.now();
+
+    const cpuUsage = await getCpuUsage();
+    
+    const totalMemory = os.totalmem();
+    const freeMemory = os.freemem();
+    const usedMemory = totalMemory - freeMemory;
+
+    const endTime = Date.now();
+    const elapsedTime = endTime - startTime;
+
+    return {
+        cpuUsage: `${cpuUsage}%`,
+        totalMemory: `${(totalMemory / (1024 * 1024)).toFixed(2)} Mo`,
+        freeMemory: `${(freeMemory / (1024 * 1024)).toFixed(2)} Mo`,
+        usedMemory: `${(usedMemory / (1024 * 1024)).toFixed(2)} Mo`,
+        elapsedTime: `${elapsedTime} ms`
+    }
+};
+
+/**
+ * Fonction permettant de retourner le taux d'utilisation du CPU.
+ * @returns Le taux d'utilisation du CPU.
+*/
+function getCpuUsage() {
+    return new Promise((resolve) => {
+        osUtils.cpuUsage((usage) => {
+            resolve((usage * 100).toFixed(2));
+        });
+    });
+};
+
 module.exports = {
     createClientOrDeliverer,
     createRestaurateur,
@@ -161,5 +206,6 @@ module.exports = {
     findUserByEmail,
     encryptPassword,
     comparePassword,
-    generateJWT
+    generateJWT,
+    getPerformanceMetrics
 };