+++ restaurant/app/controllers/restaurantController.js
@@ -1,32 +1,72 @@
-const Restaurant = require('../models/restaurantModel');
+/**
+ * Le contrôleur contenant la logique métier associée à chaque route des restaurants.
+ * @author AMARA Ahmed
+ * @version 1.0
+*/
 
-exports.createRestaurant = async (req, res) => {
-    console.log("Received request:", req.body);
+const restaurantService = require('../services/restaurantService');
+const decodeJWT = require('../utils/decodeToken');
+
+const createRestaurant = async (req, res) => {
+    if (!req.body) {
+        return res.status(400).json({ error: "Required request body is missing" });
+    }
+
+    const name = req.body["name"];
+    const address = req.body["address"];
+    const ownerID = req.body["ownerID"];
+
+    if (!name || !address || !ownerID) {
+        return res.status(400).json({ error: "Missing mandatory data" });
+    }
+
     try {
-        // Vérification des champs requis
-        if (!req.body.name || !req.body.address || !req.body.phoneNumber) {
-            console.log("Validation error: Missing required fields");
-            throw new Error('Missing required fields');
+        const existingRestaurant = await restaurantService.findRestaurant(name, ownerID, address);
+
+        if (existingRestaurant) {
+            throw new Error("Restaurant already exists");
         }
 
-        console.log("Creating new restaurant with data:", req.body);
-        const newRestaurant = new Restaurant({
-            name: req.body.name,
-            address: req.body.address,
-            phoneNumber: req.body.phoneNumber,
-            image: req.body.image || '', 
-            ownerID: req.body.ownerID || null,
-            description: req.body.description || '', 
-            status: req.body.status || false, 
-        });
+        await restaurantService.createRestaurant(name, ownerID, address);
 
-      
-        const savedRestaurant = await newRestaurant.save();
-        console.log("Restaurant saved:", savedRestaurant);
-
-        res.status(201).send(savedRestaurant);
-    } catch (err) {
-        console.log("Error in createRestaurant:", err.message);
-        res.status(400).send({ error: err.message });
+        return res.status(201).json({ message: "Restaurant created" });
     }
+    catch (error) {
+        if (error.message === "Restaurant already exists") {
+            return res.status(409).json({ error: error.message });
+        }
+        else {
+            console.error("Unexpected error while creating a restaurant :", error.message);
+            return res.status(400).send({ error: error.message });
+        }
+    }
+};
+
+const metrics = async (req, res) => {
+    const token = req.headers.authorization.split(' ')[1];
+    const userType = decodeJWT(token).type;
+
+    try {
+        if (userType != "SERVICE TECHNIQUE") {
+            throw new Error("Invalid user type");
+        }
+
+        const metrics = await restaurantService.getPerformanceMetrics();
+
+        return res.status(200).json({ metrics });
+    }
+    catch (error) {
+        if (error.message === "Invalid user type") {
+            res.status(403).json({ error: "Forbidden" });
+        }
+        else {
+            console.error("Unexpected error while getting metrics : ", error);
+            res.status(500).json({ error: "Metrics collecting failed" });
+        }
+    }
+};
+
+module.exports = {
+    createRestaurant,
+    metrics
 };