+++ restaurant/app/schemas/restaurantSchema.js
@@ -1,4 +1,12 @@
+/**
+ * Le schéma représentant un restaurant.
+ * @author GAURE Warren
+ * @version 1.0
+*/
+
 const mongoose = require('mongoose');
+
+// Définition du schéma pour un restaurant.
 const restaurantSchema = new mongoose.Schema({
     name: {
         type: String,
@@ -13,12 +21,12 @@ const restaurantSchema = new mongoose.Schema({
     ownerID: {
         type: Number,
         unique: true,
-        required: false
+        required: true
     },
     address: {
         type: String,
         unique: true,
-        required: false
+        required: true
     },
     description: {
         type: String,
@@ -28,7 +36,8 @@ const restaurantSchema = new mongoose.Schema({
     status: {
         type: Boolean,
         unique: false,
-        required: false
+        required: true,
+        default: true
     },
     menus : [{ 
         type: mongoose.Schema.Types.ObjectId,